Move mock oberserver function to own file
